Revert "Rename packages from autoconfig -> autoconfigure"
