+++ graphql-dgs-extended-scalars/src/test/kotlin/com/netflix/graphql/dgs/autoconfig/MultipleRegistrationOfExtendedScalarsTest.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.netflix.graphql.dgs.autoconfigure
+package com.netflix.graphql.dgs.autoconfig
 
 import com.netflix.graphql.dgs.*
 import graphql.scalars.ExtendedScalars
