+++ graphql-dgs-spring-boot-oss-autoconfigure/src/test/kotlin/com/netflix/graphql/dgs/autoconfig/AutoConfigWithExistingExecutableSchemaTest.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.netflix.graphql.dgs.autoconfigure
+package com.netflix.graphql.dgs.autoconfig
 
 import com.netflix.graphql.dgs.DgsQueryExecutor
 import graphql.Scalars.GraphQLString
